// tools/curl-converter/types.ts
export type CurlConverterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  targetLanguageLabel: string
  languagePlaceholder: string
  inputCurlLabel: string
  outputCodeLabel: string
  curlPlaceholder: string
  useSampleLabel: string
  clearLabel: string
  downloadCodeLabel: string
  warningsTitle: string
  errorsTitle: string
  importFromFileLabel: string
  copyResultLabel: string
  copiedLabel: string
}>